requirements.txt
beautifulsoup4==4.12.3
debugpy==1.6.2
django==4.2
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-extras==0.3
django-geoip2-extras==1.1.2
django-geojson==4.0.0
django-leaflet==0.29.0
django-raster==0.8.1
django-select2==8.2.3
django-bootstrap5==24.2
django-bootstrap-datepicker-plus == 5.0.5
drf-spectacular==0.15.1
Fiona==1.9.4.post1
gdal==3.4.1
pandas==1.5.3
geopandas==0.12.2
psycopg2==2.9.3
django-crispy-forms==2.0
crispy-bootstrap5==0.7
jsonfield==3.1.0
pytz==2022.2.1
netCDF4==1.6.3 
SQLAlchemy==2.0.9
GeoAlchemy2==0.13.2 
pycurl==7.45.2 
Js2Py==0.74 
scipy==1.10.1 
PyPrind==2.11.3 
rasterio==1.3.8
tzdata==2023.3 
geoserver-rest==2.4.1 
django-sass-processor==1.3
xmltodict==0.12.0
xarray==2024.6.0
dask[complete]
pyzmq==26.0.3
celery 
celery[redis] 
django-celery-beat
paramiko

